# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Green Shop Api Documentaion By Swagger",
        "description": "Green Shop Service Api Documentation (BehAndam Admin Services)",
        "contact": {
            "email": "saber.tabatabaee@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://debug.behaminplus.ir/shop-service",
            "description": "debug behaminplus API Server"
        },
        {
            "url": "https://develop.behaminplus.ir/shop-service",
            "description": "develop behaminplus API Server"
        },
        {
            "url": "https://behaminplus.ir/shop-service",
            "description": "stage behaminplus API Server"
        },
        {
            "url": "https://behandam.kermany.com/shop-service",
            "description": "PRODUCTION behaminplus API Server"
        },
        {
            "url": "https://shop.pardisania.ir",
            "description": "pardisania API Server"
        },
        {
            "url": "https://beitak.ir",
            "description": "beitak API Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "LOCAL API Server"
        }
    ],
    "paths": {
...
        "/api/blog-posts/{blogPostId}": {
            "get": {
                "tags": [
                    "BlogPosts"
                ],
                "summary": "Returns BlogPost by id as json",
                "description": "Returns BlogPost by id",
                "operationId": "getBlogPostById",
                "parameters": [
                    {
                        "name": "blogPostId",
                        "in": "path",
                        "description": "ID of BlogPost",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Use username & password for token",
                "name": "BearerAuth",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}